Add `DirnameExt` trait to parse data path
This will make a bit more sense in the context of `configure_me` where
we will mostly be dealing in strings. I think extension traits can be
cleaner for these types of use cases though, so I leave this as a
separate commit and build the `configure_me` changes on top.