Remove `clap` for `configure_me`
Some opinionated changes I made here:

`Network` implements `FromStr`, so I decided to make an extension
trait to map a `bitcoin::Network` to `ChainType`. This implies the
naming scheme is inherited from `bitcoin::Network`. The only change I
am aware of is `mainnet` to `bitcoin`, which I think formally
specifies the user is running on the bitcoin network.

Otherwise, this is a drop-in replacement of `clap`.