Implement server logic for IPC interface
The `RpcNetwork` is not `Send`, so it has to be run on a `LocalSet`
(basically a single-thread task manager). I have it bind to
`./node.sock` with no configuration option, but maybe this should be
tied to the data dir in the future. I have a dummy `IpcInterface` handle
the `echo` requests, but you can imagine this having a chain state,
wallet, etc in the future.