Use DNS resolution from `tokio`
It was raised to me that the `bitcoin-p2p` DNS resolution does not work
when using a VPN. Of course, this will not be acceptable for many users.
With the introduction of an IPC interface, we brought in `tokio`, so we
can leverage their implementation of DNS resolution. AFICT this resolves
the issue.