Add block stale-ness check
I noticed my home node got stuck on a block with the connection it had,
but continued to exchange ping and pongs. This adds a very basic check
to see if blocks are being withheld from us by disconnecting from a
peer if they have not sent a block to us for 20 minutes.